backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
moviepy==1.0.3
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
soundfile==0.12.1
google-generativeai==0.5.4
python-dotenv==1.1.1
elevenlabs==0.2.26
pydub==0.25.1 